externalize markdown nodes to types
Move Markdown node definitions to types, as we will be reusing them in
another file in the next commit.
